Added a fucntion to determine if a graph is bipartite.
This change adds a functions that can determine if certain graph
is bipartite. The complexity is pretty high, so further changes
will be needed.

The fucntions uses a pair of sets in which vertices of the graph
are added alternately. The first vertex is added to set 1, and all
it's adjacent vertices are added to the second. Vertices adjacent
to those last are added to set 1. Next vertices are added to set1
if they are not in set2, and the adding of adjacent vertices con-
tinues as said. If the intersection of both sets is not empty the
function returns False, True otherwise.